handle pack id for query key, refactor: sidebar
