+++ package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "@reduxjs/toolkit": "^2.0.1",
     "@tanstack/react-query": "^5.18.0",
+    "@tanstack/react-query-devtools": "^5.18.1",
     "axios": "^1.6.2",
     "chart.js": "^4.4.1",
     "hashids": "^2.3.0",
