+++ src/main.tsx
@@ -5,6 +5,7 @@ import { Provider } from 'react-redux';
 import { store } from './redux/store';
 import AppWithRoutes from './views/Layout/AppWithRoutes';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 
 const queryClient = new QueryClient();
 
@@ -14,6 +15,7 @@ ReactDOM.createRoot(document.getElementById('root')!).render(
 			<Provider store={store}>
 				<AppWithRoutes />
 			</Provider>
+			<ReactQueryDevtools initialIsOpen={false} />
 		</QueryClientProvider>
 	</>,
 );
